pyproject.toml
[tool.poetry]
name = "nos-config-collector"
version = "0.2.3"
description = ""
authors = ["Leo Kirchner <leo@kirchne.red>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = ">=0.95.1"
uvicorn = ">=0.22.0"
gitpython = ">=3.1.31"
jinja2 = ">=3.1.2"
# Until netconan gets a new release
netconan = { git = "https://github.com/intentionet/netconan", branch = "master" }
pydantic = {extras = ["dotenv"], version = "^1.10.7"}
requests = "^2.30.0"
netutils = "^1.4.1"

[tool.poetry.group.dev.dependencies]
commitizen = ">=3.2.2"
pytest = ">=7.3.1"
httpx = ">=0.24.0"
ruff = ">=0.0.267"
black = ">=23.1.0"
pytest-mock = "^3.10.0"
mypy = "^1.3.0"
requests-mock = "^1.10.0"
types-requests = "^2.31.0.0"
coverage = "^7.2.6"

[tool.ruff]
select = [
    "E",
    "F",
    "S",  # bandit
    "PL",  # pylint
    "I",  # isort
    "D",  # pydocstyle
]
ignore = [
    "E501",  # handled by Black
    "D212",
    "D203",
    "S101"
]
# Same as Black.
line-length = 88
target-version = "py37"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_type = "semver"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true

[tool.black]
line-length = 120
target-version = ['py311']

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

